requirements.txt
crewai>=0.11.0
python-dotenv>=1.0.0
langchain-openai>=0.0.5
